refactor(tests): rename and reorganize match test classes for clarity and consistency
