+++ lib/DartsScorer.x01/MatchX01Tests.cs
@@ -1,4 +1,5 @@
 using DartsScorer.Main.Match;
+using DartsScorer.Main.Match.x01;
 
 namespace DartsScorer.x01;
 
@@ -7,7 +8,7 @@ public class MatchX01Tests
     [Test]
     public void Match_X01_Instantiation()
     {
-        var match = new Main.Match.x01.Match();
+        var match = new Match();
 
         Assert.That(match.DartsMatchType, Is.EqualTo(DartsMatchType.X01));
         Assert.That(match.Players.Count, Is.EqualTo(0));
@@ -17,7 +18,7 @@ public class MatchX01Tests
     [Test]
     public void Match_X01_Instantiation_Fail()
     {
-        var match = new Main.Match.x01.Match();
+        var match = new Match();
 
         Assert.That(match.DartsMatchType, Is.Not.EqualTo(DartsMatchType.Killer));
     }
@@ -25,7 +26,7 @@ public class MatchX01Tests
     [Test]
     public void Match_X01_Instantiation_501_Success()
     {
-        var match = new Main.Match.x01.Match();
+        var match = new Match();
 
         Assert.That(match.RequiredScore, Is.EqualTo(501));
     }
@@ -33,7 +34,7 @@ public class MatchX01Tests
     [Test]
     public void Match_X01_Instantiation_Custom_Score_Success()
     {
-        var match = new Main.Match.x01.Match(1);
+        var match = new Match(1);
 
         Assert.That(match.RequiredScore, Is.EqualTo(101));
     }
@@ -41,13 +42,13 @@ public class MatchX01Tests
     [Test]
     public void Match_X01_Instantiation_Custom_Score_Must_Be_Between_One_and_Five()
     {
-        Assert.Throws<ArgumentException>(() => new Main.Match.x01.Match(12));
+        Assert.Throws<ArgumentException>(() => new Match(12));
     }
 
     [Test]
     public void Match_X01_Instantiation_Can_Add_Player()
     {
-        var match = new Main.Match.x01.Match();
+        var match = new Match();
         var player = new Main.Player.Player("Player 1");
         match.AddPlayer(player);
 
@@ -57,7 +58,7 @@ public class MatchX01Tests
     [Test]
     public void Match_X01_Instantiation_Throws_Exception_If_Match_Started_With_No_Players()
     {
-        var match = new Main.Match.x01.Match();
+        var match = new Match();
 
         Assert.Throws<InvalidOperationException>(() => match.StartMatch());
     }
@@ -65,7 +66,7 @@ public class MatchX01Tests
     [Test]
     public void Match_X01_Instantiation_Start_Match_With_Correct_Player()
     {
-        var match = new Main.Match.x01.Match();
+        var match = new Match();
         var player1 = new Main.Player.Player("Player 1");
         var player2 = new Main.Player.Player("Player 2");
         match.AddPlayer(player1);
@@ -79,7 +80,7 @@ public class MatchX01Tests
     [Test]
     public void Match_X01_Instantiation_Start_Match_With_First_Set()
     {
-        var match = new Main.Match.x01.Match();
+        var match = new Match();
         var player1 = new Main.Player.Player("Player 1");
         var player2 = new Main.Player.Player("Player 2");
         match.AddPlayer(player1);
