+++ lib/tests/DartsScore.RoundTheBoard/LegTests.cs
@@ -1,54 +1,10 @@
-using DartsScorer.Main.Match;
 using DartsScorer.Main.Match.RoundTheBoard;
 using DartsScorer.Main.Scoring;
 
 namespace DartsScore.RoundTheBoard;
 
-public class RoundTheBoardTests
+public class LegTests
 {
-    [SetUp]
-    public void Setup()
-    {
-    }
-
-    [Test]
-    public void RoundTheBoard_Instantiate()
-    {
-        var roundTheBoard = new Match();
-        
-        Assert.That(roundTheBoard.DartsMatchType, Is.EqualTo(DartsMatchType.RoundTheBoard));
-    }
-    
-    [Test]
-    public void RoundTheBoard_Instantiate_Failure()
-    {
-        var roundTheBoard = new Match();
-        
-        Assert.That(roundTheBoard.DartsMatchType, !Is.EqualTo(DartsMatchType.X01));
-        Assert.That(roundTheBoard.Players.Count, Is.EqualTo(0));
-    }
-
-    [Test]
-    public void RoundTheBoard_Can_Create_Player()
-    {
-        var roundTheBoardPlayer = new RoundTheBoardPlayer("Fancy New Player Name");
-        
-        Assert.That(roundTheBoardPlayer.RequiredBoardNumber, Is.EqualTo( 1));
-        Assert.That(roundTheBoardPlayer.Legs.Count, Is.EqualTo( 0));
-    }
-
-    [Test]
-    public void RoundTheBoard_Can_Add_Player()
-    {
-        var roundTheBoard = new Match();
-
-        var roundTheBoardPlayer = new RoundTheBoardPlayer("Fancy New Player Name");
-        
-        roundTheBoard.AddPlayer(roundTheBoardPlayer);
-        
-        Assert.That(roundTheBoard.Players.Count, Is.EqualTo(1));
-    }
-
     [Test]
     public void RoundTheBoard_Leg_Test()
     {
@@ -58,7 +14,7 @@ public class RoundTheBoardTests
         
         Assert.That(roundTheBoardPlayer.Legs.Count, Is.EqualTo(0));
     }
-    
+
     [Test]
     public void RoundTheBoard_Player_Basic_Score_Test()
     {
@@ -72,7 +28,7 @@ public class RoundTheBoardTests
         Assert.That(roundTheBoardPlayer.RequiredBoardNumber, Is.EqualTo( 2));
         Assert.That(roundTheBoardPlayer.Legs.Count, Is.EqualTo(1));
     }
-    
+
     [Test]
     public void RoundTheBoard_Player_Score_Test()
     {
@@ -85,8 +41,8 @@ public class RoundTheBoardTests
         
         Assert.That(roundTheBoardPlayer.RequiredBoardNumber, Is.EqualTo( 4));
         Assert.That(roundTheBoardPlayer.Legs.Count, Is.EqualTo(1));
-    }    
-    
+    }
+
     [Test]
     public void RoundTheBoard_Player_Multi_Leg_Score_Test()
     {
@@ -104,8 +60,8 @@ public class RoundTheBoardTests
         
         Assert.That(roundTheBoardPlayer.RequiredBoardNumber, Is.EqualTo( 4));
         Assert.That(roundTheBoardPlayer.Legs.Count, Is.EqualTo(2));
-    }   
-    
+    }
+
     [Test]
     public void RoundTheBoard_Player_Mid_Leg_Test()
     {
